refactor table with results
