+++ frontend/src/components/Tables/PropertyDetailsTableWithResults.js
