+++ frontend/src/components/Tables/PropertyDetailsTableWithoutResults.js
