+++ frontend/src/components/Tables/WithResults/WithResults.module.sass
@@ -1,0 +1,64 @@
+@import "./src/styles/helpers"
+
+.table
+  width: 100%
+  border-collapse: separate
+  border-spacing: 0
+  background-color: $white
+  border-radius: 20px
+  overflow-x: auto
+  box-shadow: 0 2px 3px 0 rgba(38, 50, 56, 0.2)
+
+  thead, tfoot
+    text-align: left
+    +body-bold-2
+
+  tbody
+    +caption-2
+
+  td, th
+    padding: 10px
+    border-top: 1px groove $n5
+
+  td
+    &:nth-child(1)
+      width: 75%
+    &:nth-child(2)
+      width: 20%
+    &:nth-child(3)
+      width: 5%
+
+  th
+    background-color: $n10
+    color: $n4
+
+  tfoot th
+    border-top: none
+    &:nth-child(2)
+      text-align: right
+      border-right: 1px groove $n5
+
+    &:first-child
+      border-bottom-left-radius: 20px
+      border-right: 1px groove $n5
+
+    &:last-child
+      border-bottom-right-radius: 20px
+      width: 50%
+
+  thead th
+    border-top: none
+
+    &:first-child
+      border-top-left-radius: 20px
+
+    &:last-child
+      border-top-right-radius: 20px
+
+  tbody td
+    &:nth-of-type(2)
+      text-align: right
+      +caption-bold-2
+      background-color: $n9
+
+
