+++ frontend/src/screens/Property/Catalog/Card/Card.module.sass
@@ -5,6 +5,7 @@
   overflow: hidden
   background: $n8
   color: $n2
+  box-shadow: 0 2px 3px 0 rgba(38, 50, 56, 0.2)
   +dark
     background: $n2
     color: $n8
