+++ frontend/src/screens/Property/Catalog/Card/Details/Details.module.sass
@@ -1,0 +1,29 @@
+@import ../../../../../styles/helpers
+
+.container
+  max-width: 1440px
+  +w
+    padding: 0 40px
+  +t
+    padding: 0
+
+.wrapper
+  padding: 80px
+  border-radius: 24px
+  background: none
+  box-shadow: inset 0 0 0 2px $white
+  +w
+    padding: 64px 40px
+  +t
+    border-radius: 0
+  +m
+    padding: 24px 32px
+    border-radius: 24px
+    box-shadow: inset 0 0 0 2px $white
+
+.title
+  margin-bottom: 64px
+  +d
+    margin-bottom: 24px
+  +m
+    text-align: center
