+++ frontend/src/screens/Property/Catalog/Card/Details/index.js
@@ -1,20 +1,20 @@
 import React from "react";
 import cn from "classnames";
-import styles from "./Details.module.sass.css";
+import styles from "./Details.module.sass";
 import useSingleProperty from "../../../../../hooks/useSingleProperty";
-import PropertyDetailsTables from "../../../../../components/tables/PropertyDetailsTables";
+import Table from "./Table";
 
 const Details = ({units}) => {
 
     const {property} = useSingleProperty()
 
     return (
-        <div className={cn("section", styles.section)}>
+        <div className={cn("section-mb64", styles.section)}>
             <div className={cn("container", styles.container)}>
                 <div className={styles.wrapper}>
-                    <h2 className={cn("h2", styles.title)}>Property details of </h2>
+                    <h2 className={cn("h2", styles.title)}>Property details of "{property.name}"</h2>
                     <div className={styles.table}>
-                        <PropertyDetailsTables property={property} units={units}/>
+                        <Table property={property} units={units}/>
                     </div>
                 </div>
             </div>
