+++ frontend/src/screens/Property/Catalog/index.js
@@ -6,7 +6,7 @@ import Card from "./Card";
 const Catalog = ({properties}) => {
 
     return (
-        <div className={cn("section", styles.section)}>
+        <div className={cn("section-mb64", styles.section)}>
             <div className={cn("container", styles.container)}>
                 <div className={styles.wrapper}>
                     <h2 className={cn("h2", styles.title)}>Your dream investments</h2>
