+++ frontend/src/styles/helpers.sass
@@ -45,7 +45,8 @@ $n2: #23262F
 $n3: #353945
 //ok -> title
 $n4: #2D3436
-$n5: #B1B5C3
+//grey 30%
+$n5: #E3E5E5
 //ok -> border
 $n6: #6C5CE7
 $n7: #F4F5F6
@@ -53,6 +54,8 @@ $n7: #F4F5F6
 $n8: #FFEAA7
 //show Value
 $n9: #FDCB6E
+//Table Header
+$n10: #ADADAD
 
 // typography
 =body-1
